+++ src/test/java/wooteco/subway/SubwayApplicationTests.java
@@ -1,4 +1,4 @@
-package wooteco.subway.acceptance;
+package wooteco.subway;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -6,8 +6,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 class SubwayApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+    @Test
+    void contextLoads() {
+    }
 
 }
