+++ src/test/java/wooteco/subway/acceptance/AcceptanceFixture.java
@@ -1,4 +1,4 @@
-package wooteco.subway.acceptance.acceptance;
+package wooteco.subway.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
