+++ src/test/java/wooteco/subway/domain/FareTest.java
@@ -1,11 +1,10 @@
-package wooteco.subway.acceptance.domain;
+package wooteco.subway.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import wooteco.subway.domain.Fare;
 
 class FareTest {
     @DisplayName("거리를 전달받아 요금을 계산한다.")
