+++ src/test/java/wooteco/subway/ui/StationControllerTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.acceptance.ui;
+package wooteco.subway.ui;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
@@ -6,9 +6,8 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-import static wooteco.subway.acceptance.Fixture.강남역;
-import static wooteco.subway.acceptance.Fixture.청계산입구역;
+import static wooteco.subway.Fixture.강남역;
+import static wooteco.subway.Fixture.청계산입구역;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
@@ -43,8 +42,8 @@ class StationControllerTest {
         final StationRequest stationRequest = new StationRequest(강남역.getName());
 
         mockMvc.perform(post("/stations")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(stationRequest)))
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(stationRequest)))
                 .andExpect(status().isCreated())
                 .andExpect(jsonPath("name").value(강남역.getName()))
                 .andDo(print());
@@ -57,8 +56,8 @@ class StationControllerTest {
         stationService.createStation(stationRequest.getName());
 
         mockMvc.perform(post("/stations")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(stationRequest)))
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(stationRequest)))
                 .andExpect(status().isBadRequest())
                 .andDo(print());
     }
